backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
passlib[bcrypt]==1.7.4
bcrypt>=3.1.0,<4.0.0
PyJWT==2.9.0
python-multipart==0.0.9
email-validator>=2.0.0

# Face Recognition
face-recognition==1.3.0
opencv-python==4.8.1.78
Pillow==10.1.0

# Geofencing
shapely==2.0.2
geopy==2.4.1
numpy<2.0.0  # Required for shapely compatibility

# Notifications
twilio==8.10.0
firebase-admin==6.3.0

# Background Tasks
APScheduler==3.10.4

